scheduler/requirements.txt
fastapi
starlette
uvicorn
gunicorn
python-multipart
requests
colorlog
pyyaml
kubernetes==20.11.0a1
